[#88, #132] balance port inputs out of order usage

+++ src/core/helper/BalancePortAnalyzer.cpp
@@ -62,7 +62,7 @@ public:
         StaticMenu(vMenu),
         vNames_(vNames),
         dig_(dig){};
-    virtual void printItem(uint8_t index) {
+    virtual void printItem(uint16_t index) {
         StaticMenu::printItem(index);
         AnalogInputs::Name name = pgm::read(&vNames_[index]);
         if(index < MAX_BANANCE_CELLS +3) {
@@ -105,7 +105,7 @@ void run() {
     Balancer::powerOn();
 
     VoltageMenu v(voltageMenu, voltageName, 5);
-    int8_t index;
+    int16_t index;
     do {
         index = v.runSimple(true);
         if(index < 1) continue;
